% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwi2fod.R
\name{dwi2fod}
\alias{dwi2fod}
\title{Estimate fibre orientation distributions
from diffusion data using spherical deconvolution}
\usage{
dwi2fod(infile, response, algorithm = c("csd", "msmt_csd"), bvals = NULL,
  bvecs = NULL, scheme = NULL, mask = NULL, lmax = NULL,
  shells = NULL, outfile = NULL, opts = "", verbose = TRUE, ...)
}
\arguments{
\item{infile}{the input DWI image containing volumes
that are both diffusion weighted and b=0}

\item{response}{response file, usually generated by \code{\link{dwi2response}}}

\item{algorithm}{the algorithm to use for fibre orientation distributions estimation}

\item{bvals}{The file containing the b-values.}

\item{bvecs}{The file containing the b-vectors.}

\item{scheme}{The file containing a 4xT text file,
instead of bvals/bvecs}

\item{mask}{initial mask for response voxel selection}

\item{lmax}{The maximum harmonic degree(s) of response function
estimation (single value for single-shell response,
comma-separated values for multi-shell response)}

\item{shells}{The b-value shell(s) to use in response
function estimation (single value for single-shell response,
comma-separated values for multi-shell response)}

\item{outfile}{Output response file}

\item{opts}{Additional options to pass}

\item{verbose}{Print diagnostic output}

\item{...}{arguments to pass to \code{\link{std_opts}}}
}
\value{
Output filename
}
\description{
Estimate fibre orientation distributions
from diffusion data using spherical deconvolution
}
\examples{
\dontrun{
infile = "~/Downloads/data.nii.gz"
bvals = "~/Downloads/bvals"
bvecs = "~/Downloads/bvecs"
response = dwi2response(infile, bvals = bvals, bvecs = bvecs)
fod = dwi2fod(infile, response = response,
bvals = bvals, bvecs = bvecs,
algorithm = "csd")
}
}
