% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwi2response.R
\name{dwi2response}
\alias{dwi2response}
\title{Estimate response function(s) for spherical deconvolution}
\usage{
dwi2response(infile, algorithm = c("tournier", "dhollander", "msmt_5tt", "fa",
  "tax", "manual"), bvals = NULL, bvecs = NULL, scheme = NULL,
  mask = NULL, lmax = NULL, shells = NULL, outfile = tempfile(),
  opts = "", verbose = TRUE)
}
\arguments{
\item{infile}{the input DWI image containing volumes
that are both diffusion weighted and b=0}

\item{algorithm}{algorithm to be used to complete
response estimation}

\item{bvals}{The file containing the b-values.}

\item{bvecs}{The file containing the b-vectors.}

\item{scheme}{The file containing a 4xT text file,
instead of bvals/bvecs}

\item{mask}{initial mask for response voxel selection}

\item{lmax}{The maximum harmonic degree(s) of response function
estimation (single value for single-shell response,
comma-separated values for multi-shell response)}

\item{shells}{The b-value shell(s) to use in response
function estimation (single value for single-shell response,
comma-separated values for multi-shell response)}

\item{outfile}{Output response file}

\item{opts}{Additional options to pass}

\item{verbose}{Print diagnostic output}
}
\value{
Output filename
}
\description{
Estimate response function(s) for spherical deconvolution
}
\examples{
\dontrun{
infile = "~/Downloads/data.nii.gz"
bvals = "~/Downloads/bvals"
bvecs = "~/Downloads/bvecs"
response = dwi2response(infile, bvals = bvals, bvecs = bvecs)
}
}
