% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_tissue_gen.R
\name{five_tissue_gen}
\alias{five_tissue_gen}
\title{Generat a five-tissue-type (5TT) segmented tissue image suitable
for use in Anatomically-Constrained Tractography (ACT)}
\usage{
five_tissue_gen(infile, algorithm = c("fsl", "freesurfer"),
  outfile = NULL, crop = TRUE, amyg_hipp_gray = FALSE, mask = NULL,
  t2 = NULL, premasked = FALSE, lut = NULL, opts = "",
  verbose = TRUE, ...)
}
\arguments{
\item{infile}{The input T1-weighted image}

\item{algorithm}{algorithm to be used to complete
segmentation}

\item{outfile}{Output image file}

\item{crop}{crop the resulting 5TT image to reduce its size (changed
dimensions compared to input image)}

\item{amyg_hipp_gray}{Represent the amygdalae and hippocampi
as sub-cortical grey matter in the 5TT image}

\item{mask}{Mask for brain}

\item{t2}{T2-weighted image in addition to the default T1-weighted image;
this will be used as a second input to FSL FAST}

\item{premasked}{Indicate that brain masking has already
been applied to the input image}

\item{lut}{Manually provide path to the lookup table on
which the input parcellation image is based (e.g. FreeSurferColorLUT.txt)}

\item{opts}{Additional options to pass to function}

\item{verbose}{Print diagnostic output}

\item{...}{arguments to pass to \code{\link{std_opts}}}
}
\value{
Output filename
}
\description{
Generat a five-tissue-type (5TT) segmented tissue image suitable
for use in Anatomically-Constrained Tractography (ACT)
}
\examples{
\dontrun{
infile = "~/Downloads/anat+seg.nii.gz"
t1 = "~/Downloads/t1.nii.gz"
response = five_tissue_gen(t1, algorithm = "freesurfer")
}
}
