% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrconvert.R
\name{mrconvert}
\alias{mrconvert}
\title{Perform conversion between different file types
and optionally extract a subset of the input image}
\usage{
mrconvert(infile, datatype = c("float32", "float32le", "float32be",
  "float64", "float64le", "float64be", "int64", "uint64", "int64le",
  "uint64le", "int64be", "uint64be", "int32", "uint32", "int32le",
  "uint32le", "int32be", "uint32be", "int16", "uint16", "int16le",
  "uint16le", "int16be", "uint16be", "cfloat32", "cfloat32le",
  "cfloat32be", "cfloat64", "cfloat64le", "cfloat64be", "int8", "uint8",
  "bit"), outfile = NULL, stride = NULL, bvals = NULL,
  bvecs = NULL, scheme = NULL, opts = "", verbose = TRUE, ...)
}
\arguments{
\item{infile}{input filename or nifti}

\item{datatype}{specify output image data type}

\item{outfile}{Output response file}

\item{stride}{specify the strides of the output data in memory,
as comma-separated values}

\item{bvals}{The file containing the b-values.}

\item{bvecs}{The file containing the b-vectors.}

\item{scheme}{The file containing a 4xT text file,
instead of bvals/bvecs}

\item{opts}{Additional options to pass}

\item{verbose}{Print diagnostic output}

\item{...}{arguments to pass to \code{\link{std_opts}}}
}
\value{
Output filename
}
\description{
Perform conversion between different file types
and optionally extract a subset of the input image
}
\examples{
\dontrun{
infile = "~/Downloads/data.nii.gz"
bvals = "~/Downloads/bvals"
bvecs = "~/Downloads/bvecs"
mrfile = mrconvert(infile, bvals = bvals, bvecs = bvecs,
outfile = tempfile(fileext = ".mif"))
}
}
