% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmath.R
\name{mrmath}
\alias{mrmath}
\title{Compute summary statistic on image intensities either across images, or
along a specified axis of a single image}
\usage{
mrmath(infile, operation = c("mean", "median", "sum", "product", "rms",
  "norm", "var", "std", "min", "max", "absmax", "magmax"),
  datatype = c("float32", "float32le", "float32be", "float64",
  "float64le", "float64be", "int64", "uint64", "int64le", "uint64le",
  "int64be", "uint64be", "int32", "uint32", "int32le", "uint32le",
  "int32be", "uint32be", "int16", "uint16", "int16le", "uint16le",
  "int16be", "uint16be", "cfloat32", "cfloat32le", "cfloat32be",
  "cfloat64", "cfloat64le", "cfloat64be", "int8", "uint8", "bit"),
  outfile = NULL, axis = NULL, opts = "", verbose = TRUE, ...)
}
\arguments{
\item{infile}{input filename or filenames.  Can be list of nifti objects too.}

\item{operation}{Supported operations are:
mean, median, sum, product, rms (root-mean-square value),
norm (vector 2-norm), var (unbiased variance),
std (unbiased standard deviation), min,
max, absmax (maximum absolute value),
magmax (value with maximum absolute
value, preserving its sign).}

\item{datatype}{specify output image data type}

\item{outfile}{Output response file}

\item{axis}{perform operation along a specified axis of a single input image}

\item{opts}{Additional options to pass}

\item{verbose}{Print diagnostic output}

\item{...}{arguments to pass to \code{\link{std_opts}}}
}
\value{
Output filename
}
\description{
Compute summary statistic on image intensities either across images, or
along a specified axis of a single image
}
\examples{
\dontrun{
infile = "~/Downloads/data.nii.gz"
}
}
