% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrtrix_bin_dir.R
\name{mrtrix_bin_dir}
\alias{mrtrix_bin_dir}
\title{Directory of Binary Executables for MRTrix}
\usage{
mrtrix_bin_dir()
}
\value{
A path
}
\description{
Directory of Binary Executables for MRTrix
}
\examples{
mrtrix_bin_dir()
file.path(mrtrix_bin_dir(), "dwi2mask")
}
