% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_opts.R
\name{std_opts}
\alias{std_opts}
\title{MRTrix Standard Options}
\usage{
std_opts(opts = "", verbose = TRUE, debug = TRUE, nthreads = 0,
  force = FALSE, cleanup = TRUE, tempdir = tempfile(),
  use_tempdir = TRUE, quiet = FALSE)
}
\arguments{
\item{opts}{Current options for the function}

\item{verbose}{Print diagnostic messages.}

\item{debug}{Display additional debugging information over and above
the verbose output}

\item{nthreads}{number of threads to use for multi-threaded operations}

\item{force}{force overwrite of output files if pre-existing}

\item{cleanup}{delete temporary files during script, or
temporary directory atscript completion}

\item{tempdir}{Manually specify the path in which to
generate the temporary directory}

\item{use_tempdir}{logical indicating if this function uses a
\code{tempdir} argument}

\item{quiet}{Suppress all console output during script execution}
}
\value{
Character string
}
\description{
MRTrix Standard Options
}
\examples{
std_opts()
}
