% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwi2mask.R
\name{dwi2mask}
\alias{dwi2mask}
\title{Transform DWI to mask}
\usage{
dwi2mask(infile, bvals = NULL, bvecs = NULL, scheme = NULL,
  outfile = NULL, clean_scale = 2, opts = "", verbose = TRUE)
}
\arguments{
\item{infile}{the input DWI image containing volumes
that are both diffusion weighted and b=0}

\item{bvals}{The file containing the b-values.}

\item{bvecs}{The file containing the b-vectors.}

\item{scheme}{The file containing a 4xT text file,
instead of bvals/bvecs}

\item{outfile}{output mask file name}

\item{opts}{Additional options to pass}

\item{verbose}{Print diagnostic output}
}
\value{
Name of output file
}
\description{
Transform DWI to mask
}
\examples{
\dontrun{
infile = "~/Downloads/data.nii.gz"
bvals = "~/Downloads/bvals"
bvecs = "~/Downloads/bvecs"
mask = dwi2mask(infile, bvals = bvals, bvecs = bvecs)
}
}
